+++ src/pages/historyPage/history.js
@@ -2,35 +2,21 @@ import { toast } from "react-toastify";
 import { Card, PagePlaceHolder, Sidebar } from "../../components";
 import { useAuth } from "../../context";
 import { useVideoGlobal } from "../../context/globalContext";
-import { deleteAllHistory } from "../../service/history/history";
+import { deleteAllHistory, HistoryApi } from "../../service/history/history";
 
 export const History = () => {
   const {
     state: { history },
-    dispatch,
   } = useVideoGlobal();
-  const {
-    auth: { token },
-  } = useAuth();
 
-  const header = { authorization: token };
-  const clearAll = async () => {
-    try {
-      const {
-        data: { history },
-      } = await deleteAllHistory(header);
-      dispatch({ type: "HISTORY", payload: history });
-    } catch (error) {
-      toast.info("something went wrong");
-    }
-  };
+  const { clearAllHistory } = HistoryApi();
 
   return (
     <div>
       <section className="main-box">
         <Sidebar />
         <main className="main-product">
-          <button className="btn" onClick={clearAll}>
+          <button className="btn" onClick={clearAllHistory}>
             Clear All History
           </button>
           {history.length === 0 && <PagePlaceHolder />}
