+++ src/service/index.js
@@ -1,3 +1,4 @@
 export { LikeVideoApi } from "./likeVideo/likeVideoApi";
 export { WatchLaterApi } from "./watchLater/watchLaterApi";
 export { addTohistory } from "./history/history";
+export { HistoryApi } from "./history/history";
